refactor(types): modify type names
