+++ src/types/AuthInfoType.tsx
